Move pipelines into their own file
