+++ src/device.cpp
@@ -1,7 +1,6 @@
 #include "device.h"
 
 #include "rhi-shared.h"
-#include "shader-object.h"
 
 #include <algorithm>
