+++ src/shader-object.cpp
@@ -1,8 +1,6 @@
 #include "shader-object.h"
 
 #include "rhi-shared.h"
-#include "device.h"
-#include "shader.h"
 
 namespace rhi {
